Refactor Image Classifier Distiller Functions
Move the majority of functionalities from compress_classifier.py to separate file imageclass.py, under distiller.
The function will henceforth will be imported automatically upon calling `import distiller`.

Also, some smaller modifications:
* evaluate_model() calls either test, or quantize and then test, which will now also calibrate.
  Calibration and testing of the quantized models will run on cloned model.
* Rename train() to train_one_epoch()
* Allow calling test and validate/test with less arguments.
  The following arguments are now optional: loggers, collectors, and args.
* acts_quant_stats_collection will return the results stats, and saving
  to file is not optional, false by default.
* qe_calibration argument is set to None, by default.
* Added get_nonparallel_clone_model to handle cloning and discarding of models.
