+++ distiller/__init__.py
@@ -26,6 +26,7 @@ from .policy import *
 from .thinning import *
 from .knowledge_distillation import KnowledgeDistillationPolicy, DistillationLossWeights
 from .summary_graph import SummaryGraph, onnx_name_2_pytorch_name
+import imageclass
 
 import logging
 logging.captureWarnings(True)
