+++ distiller/utils.py
@@ -21,6 +21,7 @@ with some random helper functions.
 """
 import argparse
 from collections import OrderedDict
+import contextlib
 from copy import deepcopy
 import logging
 import operator
@@ -646,6 +647,15 @@ def make_non_parallel_copy(model):
     replace_data_parallel(new_model)
 
     return new_model
+
+
+@contextlib.contextmanager
+def get_nonparallel_clone_model(model):
+    clone_model = make_non_parallel_copy(model)
+    try:
+        yield clone_model
+    finally:
+        del clone_model
 
 
 def set_seed(seed):
